+++ sdk/identity/identity-cache-persistence/samples-dev/basic.ts
@@ -8,6 +8,10 @@
  * `DeviceCodeCredential`, will be able to retrieve tokens from the cache rather
  * than requesting new tokens from the Azure Active Directory token endpoint.
  *
+ * In order to utilize the persistent token cache, the `enabled` property must
+ * be set to `true` within `tokenCachePersistenceOptions` in the credential's
+ * options.
+ *
  * @summary import and use the persistence extension
  */
 
@@ -20,7 +24,13 @@ import dotenv from "dotenv";
 dotenv.config();
 
 async function main() {
-  const credential = new DeviceCodeCredential();
+  const credential = new DeviceCodeCredential({
+    // This property must be provided, with `enabled` set to true to enable
+    // persistent token caching.
+    tokenCachePersistenceOptions: {
+      enabled: true
+    }
+  });
 
   // This is the scope we will use to get a token from the AAD token endpoint.
   // By default, we'll use the Microsoft Graph scope as an example, but when
@@ -29,9 +39,11 @@ async function main() {
   const scope = process.env.AAD_TEST_SCOPE ?? "https://graph.microsoft.com/.default";
 
   // A little helper function to print an access_token
-  const logToken = async () => console.log(await (await credential.getToken(scope)).token);
+  const logToken = async () => console.log((await credential.getToken(scope)).token);
 
-  // You should observe that the same token is printed twice. if not, then caching isn't working properly.
+  // You should observe that the same token is printed twice, and that running the sample twice
+  // only prompts for interactive authentication once. In the second execution, the token should
+  // be retrieved from the cache.
   console.log("Calling getToken()", 1);
   await logToken();
   console.log("Calling getToken()", 2);