+++ sdk/identity/identity-cache-persistence/src/index.ts
@@ -1,18 +1,42 @@
-// Copyright (c) Microsoft Corporation
+// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-const persistence: unique symbol = Symbol("identity-cache-persistence");
-declare module "@azure/identity" {
-  export interface AzureIdentityExtensionTypeMap {
-    [persistence]: "@azure/identity-cache-persistence";
-  }
-}
+import { AzureExtensionContext, IdentityExtension } from "../../identity/src/extensions/provider";
+import { createPersistenceCachePlugin } from "./provider";
 
-import { registerExtension } from "../../identity/src/extensions/provider";
-import { createPersistenceCachePlugin } from "./persistence";
+/**
+ * An extension that provides persistent token caching for `@azure/identity`
+ * credentials. The extension API is compatible with `@azure/identity` versions
+ * 2.0.0 and later. Load this extension using the `useIdentityExtension`
+ * function, imported from `@azure/identity`.
+ *
+ * In order to enable this functionality, you must also pass
+ * `tokenCachePersistenceOptions` to your credential constructors with an
+ * `enabled` property set to true.
+ *
+ * Example:
+ *
+ * ```typescript
+ * import { useIdentityExtension, DeviceCodeCredential } from "@azure/identity";
+ * import persistence from "@azure/identity-cache-persistence";
+ *
+ * // Load the extension
+ * useIdentityExtension(persistence);
+ *
+ * const credential = new DeviceCodeCredential({
+ *   tokenCachePersistenceOptions: {
+ *     enabled: true
+ *   }
+ * });
+ * ```
+ */
 
-registerExtension(persistence, ({ cachePluginControl }) => {
+const persistence: IdentityExtension = (context) => {
+  const { cachePluginControl } = context as AzureExtensionContext;
+
   cachePluginControl.persistence = createPersistenceCachePlugin;
-});
+};
 
 export default persistence;
+
+export { IdentityExtension };