+++ sdk/identity/identity-cache-persistence/test/internal/node/setup.spec.ts
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Copyright (c) Microsoft Corporation
+// Licensed under the MIT license.
+
+// We need to set up the extension for the tests!
+
+import { useIdentityExtension } from "../../../../identity/src";
+import extension from "../../../src";
+
+before(function() {
+  useIdentityExtension(extension);
+});